Current progress of the environment implementation
We need to reevaluate the code structure for the combat scene. Delegating specific tasks to areas where it is very flexible for future implementations. For instance, something we will be thinking about later is Status, delegating status to a specific class branched off from the combatscenemanager is necessary to be readable. Thereby also passing over needed information from the combat scene manager