Fixed a whole lot of systems.
We have made it so that the combatComponent is the only place where combat data is stored. Other components can read this data but it isn't to be edited from else where.
We have also made a combat Attack handler class which takes that data to enact attack data. This class is the means for which the AttackActions list in resolution data is edited. Specifically, the list's 0 entry is removed to allow for the next move to be enacted when the time comes